test: rename and move contracts
