+++ src/test/suite/cairo-parser/getScopeParsingResult.test.ts
@@ -6,7 +6,7 @@ suite("get-scope-parsing-result", () => {
   test("constructor", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/ERC20.cairo"
+      "../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
 
     // parse whole scope
@@ -85,7 +85,7 @@ suite("get-scope-parsing-result", () => {
   test("view", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/ERC20.cairo"
+      "../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
 
     // parse whole scope
@@ -308,7 +308,7 @@ suite("get-scope-parsing-result", () => {
   test("external", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/ERC20.cairo"
+      "../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
 
     // parse whole scope
