+++ src/test/suite/function-comment/constructor/returns.test.ts
@@ -7,7 +7,7 @@ suite("function-comment: constructor: returns", () => {
   test("parse line 12", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../../testAssets/ERC20.cairo"
+      "../../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
 
     const functionText = CairoParser.parseFunctionScope(
@@ -48,7 +48,7 @@ suite("function-comment: constructor: returns", () => {
   test("parse line 13", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../../testAssets/ERC20.cairo"
+      "../../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
     const functionText = CairoParser.parseFunctionScope(
       pathFile,
@@ -84,7 +84,7 @@ suite("function-comment: constructor: returns", () => {
   test("parse line 14", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../../testAssets/ERC20.cairo"
+      "../../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
     const functionText = CairoParser.parseFunctionScope(
       pathFile,
@@ -123,7 +123,7 @@ suite("function-comment: constructor: returns", () => {
   test("parse whole scope", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../../testAssets/ERC20.cairo"
+      "../../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
     const functionText = CairoParser.parseFunctionScope(
       pathFile,
