+++ src/test/suite/function-scope/constructor.test.ts
@@ -12,7 +12,7 @@ suite("integration-test: constructor", () => {
   test("0", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/ERC20.cairo"
+      "../../../../testContracts/ERC20Compliant/ERC20.cairo"
     );
 
     // parse whole scope
