+++ src/test/suite/function-scope/storage_var.test.ts
@@ -12,7 +12,7 @@ suite("integration-test: event", () => {
   test("0", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
@@ -96,7 +96,7 @@ suite("integration-test: event", () => {
   test("1", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
 
@@ -708,7 +708,7 @@ suite("integration-test: event", () => {
   test("2", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
@@ -807,7 +807,7 @@ suite("integration-test: event", () => {
   test("3", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
@@ -906,7 +906,7 @@ suite("integration-test: event", () => {
   test("4", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
@@ -1011,7 +1011,7 @@ suite("integration-test: event", () => {
   test("5", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
@@ -1120,7 +1120,7 @@ suite("integration-test: event", () => {
   test("6", () => {
     const pathFile = path.resolve(
       __dirname,
-      "../../../../testAssets/library.cairo"
+      "../../../../testContracts/ERC20Compliant/library.cairo"
     );
 
     // parse whole scope
